third_party_model/anysplat/requirements.txt
numpy==1.25.0
wheel
tqdm
lightning
black
ruff
hydra-core
jaxtyping
beartype
wandb
einops
colorama
scikit-image
colorspacious
matplotlib
moviepy
imageio
timm
dacite
lpips
e3nn
plyfile
tabulate
svg.py
scikit-video
opencv-python
Pillow
huggingface_hub
gradio
xformers==0.0.24
torch_scatter==2.1.2
moviepy==1.0.3
pydantic
open3d
einops
safetensors
git+https://github.com/facebookresearch/pytorch3d.git
https://github.com/nerfstudio-project/gsplat/releases/download/v1.4.0/gsplat-1.4.0%2Bpt22cu121-cp310-cp310-linux_x86_64.whl
